+++ src/app_v2/main.py
@@ -1,0 +1,9 @@
+from app import SSApp
+
+if __name__ == "__main__":
+    app = SSApp()
+    try:
+        app.run()
+    except KeyboardInterrupt:
+        print()
+        exit()
